refactor: Update pointage_travaux interface and add pointage_travauxStore interface
This commit refactors the 'pointage_travaux' interface by adding a new property 'metre_travaux' to represent the measurement of work done. Additionally, a new interface 'pointage_travauxStore' is introduced to store the pointage data along with a message and selected ID. These changes enhance the functionality and data management of the application.